Add files via upload
# Dogs-vs.-Cats-Prediction-Problem
Develop a Deep Convolutional Neural Network Step-by-Step to Classify Photographs of Dogs and Cats
The Dogs vs. Cats dataset is a standard computer vision dataset that involves classifying photos as either containing a dog or cat.

Although the problem sounds simple, it was only effectively addressed in the last few years using deep learning convolutional neural networks. While the dataset is effectively solved, it can be used as the basis for learning and practicing how to develop, evaluate, and use convolutional deep learning neural networks for image classification from scratch.

This includes how to develop a robust test harness for estimating the performance of the model, how to explore improvements to the model, and how to save the model and later load it to make predictions on new data.

In this tutorial, you will discover how to develop a convolutional neural network to classify photos of dogs and cats.

After completing this , you will know:

How to load and prepare photos of dogs and cats for modeling.
How to develop a convolutional neural network for photo classification from scratch and improve model performance.
How to develop a model for photo classification using transfer learning.
Dogs vs. Cats Prediction Problem
Dogs vs. Cats Dataset Preparation
Develop a Baseline CNN Model
Develop Model Improvements
Explore Transfer Learning
How to Finalize the Model and Make Predictions